add exception handler in each microservice with exceptions

+++ backend/synthexITBackend/employer-service/src/main/java/com/devsling/fr/service/EmployerServiceImpl.java
@@ -1,9 +1,10 @@
 package com.devsling.fr.service;
 
 import com.devsling.fr.dto.EmployerDto;
+import com.devsling.fr.exceptions.EmployerException;
 import com.devsling.fr.repository.EmployerRepository;
-import com.devsling.fr.service.EmployerService;
 import com.devsling.fr.tools.AppUtils;
+import com.devsling.fr.tools.Constants;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import reactor.core.publisher.Flux;
@@ -17,11 +18,15 @@ public class EmployerServiceImpl implements EmployerService {
 
     @Override
     public Flux<EmployerDto> getEmployer(){
-        return employerRepository.findAll().map(AppUtils::entityToDto);
+        return employerRepository.findAll().map(AppUtils::entityToDto)
+                .switchIfEmpty(Mono.error(new EmployerException(Constants.NO_EMPLOYER_FOUND)));
+
     }
     @Override
-    public Mono<EmployerDto> getEmployer(Long id){
-        return employerRepository.findById(id).map(AppUtils::entityToDto);
+    public Mono<EmployerDto> getEmployerById(Long id){
+        return employerRepository.findById(id).map(AppUtils::entityToDto)
+                .switchIfEmpty(Mono.error(new EmployerException(Constants.NO_EMPLOYER_FOUND_WITH_ID_+id)));
+
     }
 
     @Override
@@ -29,7 +34,8 @@ public class EmployerServiceImpl implements EmployerService {
         return candidateDtoMono
                 .map(AppUtils::dtoToEntity)
                 .flatMap(employerRepository::save)
-                .map(AppUtils::entityToDto);
+                .map(AppUtils::entityToDto)
+                .switchIfEmpty(Mono.error(new EmployerException(Constants.SAVE_EMPLOYER_ERROR_MESSAGE)));
     }
     @Override
     public Mono<EmployerDto> updateEmployer(Mono<EmployerDto> employerDtoMono, Long id){
@@ -37,17 +43,29 @@ public class EmployerServiceImpl implements EmployerService {
                 .flatMap(p->employerDtoMono.map(AppUtils::dtoToEntity)
                         .doOnNext(e->e.setId(id)))
                 .flatMap(employerRepository::save)
-                .map(AppUtils::entityToDto);
+                .map(AppUtils::entityToDto)
+                .switchIfEmpty(Mono.error(new EmployerException(Constants.UPDATE_EMPLOYER_ERROR_MESSAGE)));
+
 
     }
     @Override
-    public Mono<Void> deleteEmployer(Long id){
-        return employerRepository.deleteById(id);
+    public Mono<Void> deleteEmployerById(Long id){
+        return employerRepository.findById(id)
+                .flatMap(employer -> {
+                    if (employer != null) {
+                        return employerRepository.deleteById(id);
+                    } else {
+                        return Mono.error(new EmployerException(Constants.DELETE_EMPLOYER_ERROR_MESSAGE));
+                    }
+                });
+
     }
 
     @Override
     public Mono<EmployerDto> getEmployerByEmail(String email) {
-        return employerRepository.findByEmail(email).map(AppUtils::entityToDto);
+        return employerRepository.findByEmail(email).map(AppUtils::entityToDto)
+                .switchIfEmpty(Mono.error(new EmployerException(Constants.NO_EMPLOYER_WITH_EMAIL_FOUND)));
+
     }
 
 }
